requirements.txt
sounddevice>=0.5.0
numpy>=1.24
PyQt6>=6.5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
